requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
pymupdf>=1.23.0
unstructured>=0.11.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
replicate>=0.22.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
requests>=2.31.0 